backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
pytest==7.4.3
httpx==0.25.1
alembic==1.12.1
requests==2.31.0
huggingface-hub[cli,inference]>=0.21.0
redis==5.0.1 